Se Solucionan errores es tipo de datos, y se minimaliza el home

+++ Domotica/Librerias/displayLCD.h
@@ -29,16 +29,16 @@ class display : public LiquidCrystal_PCF8574{
 		void brillo(uint32_t _brilloV);
 		void encender(uint8_t como);
 		void apagar(uint8_t como);
-		void imprimir(uint8_t x, uint8_t y, String dato);
-		void imprimir(uint8_t x, uint8_t y, int dato);
+		void imprimir(int x, int y, String dato);
+		void imprimir(int x, int y, int dato);
 		void gotoxy(int x, int y);
-		void printFecha( uint8_t x, uint8_t y,int dia, int dia_semana, int mes, int anio);
-		void printHora(uint8_t x, uint8_t y, int hora, int minutos, int segundos);
-		void printSimbolo(uint8_t x, uint8_t y, int simbolo);
+		void printFecha(int x, int y,int dia, int dia_semana, int mes, int anio);
+		void printHora(int x, int y, int hora, int minutos, int segundos);
+		void printSimbolo(int x, int y, int simbolo);
 		void reset();
-		void printClima(uint8_t x, uint8_t y, uint8_t valor, uint8_t valorF, int tipoClima);
-		void printBT(uint8_t x, uint8_t y, bool BT_ESTADO);
-		void printWiFi(uint8_t x, uint8_t y, bool WIFI_ESTADO);
+		void printClima(int x, int y, int valor, int valorF, int tipoClima);
+		void printBT(int x, int y, bool BT_ESTADO);
+		void printWiFi(int x, int y, bool WIFI_ESTADO);
 	
 };
